Less files, same functionality, better reading
