--- server/src/core/apps/UseCases/CheckIfAppExists.js
@@ -1,10 +1,0 @@
-import { AppDoesNotExist } from '../Errors'
-
-export default (
-  appsService
-) => async (ctx) => {
-  if (await appsService.exist(ctx.app))
-    return ctx
-  else
-    throw new AppDoesNotExist('App does not exist')
-}
