--- server/src/core/apps/UseCases/ListAllApps.js
@@ -1,9 +1,0 @@
-const ListAllApps = (
-  appsService
-) => async (ctx) => {
-  const appList = await appsService.list()
-
-  return { ...ctx, apps: appList }
-}
-
-export default ListAllApps
