--- server/src/core/apps/features/ListApps.js
@@ -1,9 +1,0 @@
-import ListAllApps from '../UseCases/ListAllApps'
-
-const ListApps = (
-  appsService,
-) => () =>
-  Promise.resolve({})
-         .then(ListAllApps(appsService))
-
-export default ListApps
