--- server/src/core/apps/features/ShowStatus.js
@@ -1,10 +1,0 @@
-import AllAppsStatus from '../UseCases/AllAppsStatus'
-
-const ShowStatus = (
-  appsService,
-  messages,
-) => () =>
-  Promise.resolve({})
-         .then(AllAppsStatus(appsService, messages))
-
-export default ShowStatus
