+++ server/src/core/features.js
@@ -1,12 +1,15 @@
-import ListApps from './apps/features/ListApps'
-import AddApp from './apps/features/AddApp'
-import RemoveApp from './apps/features/RemoveApp'
-import ShowStatus from './apps/features/ShowStatus'
-import TakeApp from './apps/features/TakeApp'
-import ReturnApp from './apps/features/ReturnApp'
 import SetReminder from './reminders/features/SetReminder'
 import CancelReminder from './reminders/features/CancelReminder'
 import NotifyTeam from './notifications/features/NotifyTeam'
+
+import {
+  ListApps,
+  AddApp,
+  RemoveApp,
+  ShowStatus,
+  TakeApp,
+  ReturnApp,
+} from './apps/UseCases'
 
 const takeApp = (
   appsService,
