--- server/test/core/apps/UseCases/CheckIfAppExists.spec.js
@@ -1,22 +1,0 @@
-import CheckIfAppExists from '#/src/core/apps/UseCases/CheckIfAppExists'
-import { AppDoesNotExist } from '#/src/core/apps/Errors'
-
-describe('CheckIfAppExists', () => {
-  const mockedAppRepo = {
-    exist: (app) => app == 'appB' ? Promise.resolve(true) : Promise.resolve(false)
-  }
-  const checkIfAppExists = CheckIfAppExists(mockedAppRepo)
-
-  it('when app exists, returns the context given', async () => {
-    const originalCtx = { app: 'appB', something: 'else' }
-    const ctx = await checkIfAppExists(originalCtx)
-    expect(ctx).toEqual(originalCtx)
-  })
-
-  it('when app does not exist, fails with error message', (done) => {
-    checkIfAppExists({ app: 'appC' }).catch((error) => {
-      expect(error instanceof AppDoesNotExist).toBeTruthy()
-      done()
-    })
-  })
-})
