--- server/test/core/apps/features/AddApp.spec.js
@@ -1,40 +1,0 @@
-import Context from '#/test/support/TestContext'
-import AddApp from '#/src/core/apps/features/AddApp'
-import { AppAlreadyExists } from '#/src/core/apps/Errors'
-
-const arrayContaining = expect.arrayContaining
-
-describe('AddApp', () => {
-  let addApp = null
-
-  beforeEach(async () => {
-    await Context.reset()
-
-    addApp = Context.buildFn(AddApp)
-  })
-
-  afterAll(async () => {
-    await Context.reset()
-    Context.exit()
-  })
-
-  it('adds an app to the app list', async () => {
-    await addApp({app: {name: 'appA'}})
-    const apps = await Context.appsService.list()
-
-    expect(apps).toEqual(arrayContaining([
-      { id: 'appA', name: 'appA' }
-    ]))
-  })
-
-  it('does not add an existing app', async () => {
-    await Context.appsService.add({ name: 'appA' })
-
-    try {
-      await addApp({app: {name: 'appA'}})
-      fail('Expected to fail')
-    } catch(error) {
-      expect(error).toBeInstanceOf(AppAlreadyExists)
-    }
-  })
-})
